requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langextract==1.0.8
pydantic==2.5.0
aiofiles==23.2.0

# File processing dependencies
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.1.0
pytesseract==0.3.10

# Optional: For enhanced PDF processing
pymupdf==1.23.8
pydantic-settings
